backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
opencv-python==4.8.1.78
easyocr==1.7.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pillow>=10.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2